Extract preconnected point grouping
